requirements.txt
# requirements.txt (generated from environment.yml)
# Note: PyTorch / CUDA not included — install separately if needed

numpy
scipy
pandas
Pillow
tqdm
scikit-learn

absl-py==2.3.1
cloudpickle==3.1.2
contourpy==1.1.1
cycler==0.12.1
etils==1.3.0
farama-notifications==0.0.4
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
glfw==2.10.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.0.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
kiwisolver==1.4.7
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mujoco==3.2.3
networkx==3.1
packaging==26.0
pyopengl==3.1.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyyaml==6.0.3
six==1.17.0
stable-baselines3==2.4.1
sympy==1.13.3
tbb==2021.13.1
typing-extensions==4.13.2
tzdata==2025.3
zipp==3.20.2
